fuse: remove trailing \0 more efficiently.

+++ fs/dirstream_darwin.go
@@ -1,11 +1,11 @@
-//go:build !darwin
-
 // Copyright 2019 the Go-FUSE Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
 package fs
 
-func getdents(fd uintptr, buf []byte) (int, error) {
+import "golang.org/x/sys/unix"
+
+func getdents(fd int, buf []byte) (int, error) {
 	return unix.Getdirentries(fd, buf, nil)
 }
